% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REGModel.R
\name{REGModel}
\alias{REGModel}
\title{R6 class representing a regression model}
\description{
Contains fields storing data and methods to build, process and visualize
a regression model.
}
\examples{
library(survival)
test1 <- data.frame(
  time = c(4, 3, 1, 1, 2, 2, 3),
  status = c(1, 1, 1, 0, 1, 1, 0),
  x = c(0, 2, 1, 1, 1, 0, 0),
  sex = c(0, 0, 0, 0, 1, 1, 1)
)
mm <- REGModel$new(as.data.frame(test1), Surv(time, status) ~ x + strata(sex), f = survival::coxph)
mm
as.data.frame(mm$result)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{a \code{data.table} storing modeling data.}

\item{\code{recipe}}{an R \code{formula} storing model formula.}

\item{\code{f}}{an R \code{function} storing model constructor, e.g., \link[survival:coxph]{survival::coxph}.}

\item{\code{focal_term}}{a user focused term (covariable).}

\item{\code{terms}}{all terms (covariables, i.e. columns) used for building model.}

\item{\code{args}}{other arguments used for building model.}

\item{\code{model}}{a constructed model.}

\item{\code{type}}{model type (class).}

\item{\code{result}}{model result, a object of \code{parameters_model}. Can be converted into
data.frame with \code{\link[=as.data.frame]{as.data.frame()}} or \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{REGModel$new()}}
\item \href{#method-plot}{\code{REGModel$plot()}}
\item \href{#method-print}{\code{REGModel$print()}}
\item \href{#method-clone}{\code{REGModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Build a \code{REGModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REGModel$new(data, recipe, f, ..., exp = TRUE, ci = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{data.table} storing modeling data.}

\item{\code{recipe}}{an R \code{formula} storing model formula.}

\item{\code{f}}{a R \code{function} storing model constructor, e.g., \link[survival:coxph]{survival::coxph}.}

\item{\code{...}}{other parameters passing to corresponding regression model function.}

\item{\code{exp}}{indicating whether or not to exponentiate the the coefficients.}

\item{\code{ci}}{confidence Interval (CI) level. Default to 0.95 (95\%).
e.g. \code{\link[survival:coxph]{survival::coxph()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{REGModel} R6 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
print the \code{REGModel} result with default plot methods from \strong{parameters} package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REGModel$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print the \code{REGModel} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REGModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REGModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
