% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regport-package.R
\docType{package}
\name{regport-package}
\alias{regport}
\alias{regport-package}
\title{regport: Regression Model Processing Port}
\description{
Provides R6 classes, methods and utilities to construct, analyze, summarize, and visualize regression models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShixiangWang/regport}
  \item \url{https://shixiangwang.github.io/regport/}
  \item Report bugs at \url{https://github.com/ShixiangWang/regport/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
